// src/app/models/test-model.ts
export class TestModel {

  id: any;

  categoryName: string;

  name: string;

  price: number;

  code: string;

  quantity: number;

  total: number;

  currencyCode: string;

  fileNumber: string;

  lastName: string;

  firstName: string;

  otherName: string;

  phoneNumber: string;

  selectedPhoneNumber: any;

  isInValid: boolean;

  randomCode: string;
}